+++ client/src/features/books/components/Card/Info/Image.tsx
